add additional check to make sure there is a parent

+++ packages/home_index/main.py
@@ -698,7 +698,7 @@ def update_metadata(
             shutil.rmtree(by_id_path)
         if by_path_path.is_symlink():
             by_path_path.unlink()
-        if by_path_path.parent != BY_PATH_DIRECTORY:
+        if by_path_path.parent and by_path_path.parent != BY_PATH_DIRECTORY:
             total_count = len(list(by_path_path.parent.iterdir()))
             if total_count == 0 and by_path_path.parent.exists():
                 shutil.rmtree(by_path_path.parent)
